- implement PR `#1151` - remove and merge `Hooks.Firing.Speed.cpp` - add `SelfHealing_CombatDelay` for warhead - make `ShapeTextPrinter::GetSignIndex` function more readable - update `Memory.h`

+++ src/Utilities/ShapeTextPrinter.h
@@ -24,7 +24,7 @@ public:
 class ShapeTextPrinter
 {
 private:
-	static int GetSignIndex(const char sign);
+	static size_t GetSignIndex(const char sign);
 	static std::vector<int> BuildFrames(const std::string& text, const int baseNumberFrame, const int baseSignFrame);
 
 public:
